18680 - logic error in iptables agent
The manager used to raise on any error - so it would be impossible to
determine the blocked? status of an ip address based on its result as it
would only reply with it on success.

This commit changes the signature of the block and unblock methods to
return a 3 item array with success, status and blocked status